refactor(core): move GlobalState and EntryStatus to types module
