+++ crates/core/src/rpc/mod.rs
@@ -74,7 +74,7 @@ impl State for Option<RunloopContext> {
 
 impl Metadata for RunloopContext {}
 
-use crate::simnet::GlobalState;
+use crate::types::GlobalState;
 use jsonrpc_core::futures::FutureExt;
 use std::future::Future;
 use surfpool_types::{
