+++ crates/core/src/types.rs
@@ -1,180 +1,166 @@
-use chrono::{DateTime, Local};
-use crossbeam_channel::{Receiver, Sender};
-use litesvm::types::TransactionMetadata;
-use solana_client::rpc_config::RpcSendTransactionConfig;
+use base64::prelude::{Engine, BASE64_STANDARD};
+use litesvm::{types::TransactionMetadata, LiteSVM};
+use solana_client::{nonblocking::rpc_client::RpcClient, rpc_response::RpcPerfSample};
 use solana_sdk::{
-    blake3::Hash,
-    clock::Clock,
     epoch_info::EpochInfo,
-    pubkey::Pubkey,
-    transaction::{TransactionError, VersionedTransaction},
+    signature::Signature,
+    transaction::{Transaction, TransactionError, TransactionVersion},
 };
-use solana_transaction_status::TransactionConfirmationStatus;
-use std::{collections::HashMap, path::PathBuf};
-use txtx_addon_network_svm::codec::subgraph::{PluginConfig, SubgraphRequest};
-use uuid::Uuid;
+use solana_transaction_status::{
+    option_serializer::OptionSerializer, EncodedConfirmedTransactionWithStatusMeta,
+    EncodedTransaction, EncodedTransactionWithStatusMeta, TransactionConfirmationStatus,
+    TransactionStatus, UiCompiledInstruction, UiInnerInstructions, UiInstruction, UiMessage,
+    UiRawMessage, UiReturnDataEncoding, UiTransaction, UiTransactionReturnData,
+    UiTransactionStatusMeta,
+};
+use std::{
+    collections::{HashMap, VecDeque},
+    sync::Arc,
+    time::{SystemTime, UNIX_EPOCH},
+};
 
-pub const DEFAULT_RPC_URL: &str = "https://api.mainnet-beta.solana.com";
-
-#[derive(Clone, Debug, PartialEq, Eq, Default)]
-pub enum RunloopTriggerMode {
-    #[default]
-    Clock,
-    Manual,
-    Transaction,
+pub struct GlobalState {
+    pub svm: LiteSVM,
+    pub transactions: HashMap<Signature, EntryStatus>,
+    pub perf_samples: VecDeque<RpcPerfSample>,
+    pub transactions_processed: u64,
+    pub epoch_info: EpochInfo,
+    pub rpc_client: Arc<RpcClient>,
 }
 
-#[derive(Debug, Clone)]
-pub struct Collection {
-    pub uuid: Uuid,
-    pub name: String,
-    pub entries: Vec<Entry>,
-}
-
-#[derive(Debug, Clone)]
-pub struct Entry {
-    pub uuid: Uuid,
-    pub values: HashMap<String, String>,
-}
-
-#[derive(Debug)]
-pub enum SubgraphEvent {
-    EndpointReady,
-    InfoLog(DateTime<Local>, String),
-    ErrorLog(DateTime<Local>, String),
-    WarnLog(DateTime<Local>, String),
-    DebugLog(DateTime<Local>, String),
-    Shutdown,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub enum SchemaDataSourcingEvent {
-    Rountrip(Uuid),
-    ApplyEntry(Uuid, String), //, SubgraphRequest, u64),
-}
-
-#[derive(Debug, Clone)]
-pub enum SubgraphCommand {
-    CreateSubgraph(Uuid, SubgraphRequest, Sender<String>),
-    ObserveSubgraph(Receiver<SchemaDataSourcingEvent>),
-    Shutdown,
-}
-
-#[derive(Debug)]
-pub enum SimnetEvent {
-    Ready,
-    Aborted(String),
-    Shutdown,
-    ClockUpdate(Clock),
-    EpochInfoUpdate(EpochInfo),
-    BlockHashExpired,
-    InfoLog(DateTime<Local>, String),
-    ErrorLog(DateTime<Local>, String),
-    WarnLog(DateTime<Local>, String),
-    DebugLog(DateTime<Local>, String),
-    PluginLoaded(String),
-    TransactionReceived(DateTime<Local>, VersionedTransaction),
-    TransactionProcessed(
-        DateTime<Local>,
-        TransactionMetadata,
-        Option<TransactionError>,
-    ),
-    AccountUpdate(DateTime<Local>, Pubkey),
-}
-
-pub enum TransactionStatusEvent {
-    Success(TransactionConfirmationStatus),
-    SimulationFailure((TransactionError, TransactionMetadata)),
-    ExecutionFailure((TransactionError, TransactionMetadata)),
-}
-
-pub enum SimnetCommand {
-    SlotForward,
-    SlotBackward,
-    UpdateClock(ClockCommand),
-    UpdateRunloopMode(RunloopTriggerMode),
-    TransactionReceived(
-        Hash,
-        VersionedTransaction,
-        Sender<TransactionStatusEvent>,
-        RpcSendTransactionConfig,
-    ),
-}
-
-pub enum PluginManagerCommand {
-    LoadConfig(Uuid, PluginConfig, Sender<String>),
-}
-
-pub enum ClockCommand {
-    Pause,
-    Resume,
-    Toggle,
-    UpdateSlotInterval(u64),
-}
-
-pub enum ClockEvent {
-    Tick,
-    ExpireBlockHash,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct SurfpoolConfig {
-    pub simnet: SimnetConfig,
-    pub rpc: RpcConfig,
-    pub subgraph: SubgraphConfig,
-    pub plugin_config_path: Vec<PathBuf>,
-}
-
-#[derive(Clone, Debug)]
-pub struct SimnetConfig {
-    pub remote_rpc_url: String,
-    pub slot_time: u64,
-    pub runloop_trigger_mode: RunloopTriggerMode,
-    pub airdrop_addresses: Vec<Pubkey>,
-    pub airdrop_token_amount: u64,
-}
-
-impl Default for SimnetConfig {
-    fn default() -> Self {
+impl GlobalState {
+    pub fn new(svm: LiteSVM, epoch_info: &EpochInfo, rpc_client: Arc<RpcClient>) -> Self {
         Self {
-            remote_rpc_url: DEFAULT_RPC_URL.to_string(),
-            slot_time: 0,
-            runloop_trigger_mode: RunloopTriggerMode::Clock,
-            airdrop_addresses: vec![],
-            airdrop_token_amount: 0,
+            svm,
+            transactions: HashMap::new(),
+            perf_samples: VecDeque::new(),
+            transactions_processed: 0,
+            epoch_info: epoch_info.clone(),
+            rpc_client,
         }
     }
 }
 
-#[derive(Clone, Debug, Default)]
-pub struct SubgraphConfig {}
-
-#[derive(Clone, Debug)]
-pub struct RpcConfig {
-    pub bind_host: String,
-    pub bind_port: u16,
-    pub remote_rpc_url: String,
+#[derive(Debug, Clone)]
+pub enum EntryStatus {
+    Received,
+    Processed(TransactionWithStatusMeta),
 }
 
-impl RpcConfig {
-    pub fn get_socket_address(&self) -> String {
-        format!("{}:{}", self.bind_host, self.bind_port)
+impl EntryStatus {
+    pub fn expect_processed(&self) -> &TransactionWithStatusMeta {
+        match &self {
+            EntryStatus::Received => unreachable!(),
+            EntryStatus::Processed(status) => status,
+        }
     }
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct SubgraphPluginConfig {
-    pub uuid: Uuid,
-    pub ipc_token: String,
-    pub subgraph_request: SubgraphRequest,
+#[derive(Debug, Clone)]
+pub struct TransactionWithStatusMeta(
+    pub u64,
+    pub Transaction,
+    pub TransactionMetadata,
+    pub Option<TransactionError>,
+);
+
+impl TransactionWithStatusMeta {
+    pub fn into_status(&self, current_slot: u64) -> TransactionStatus {
+        TransactionStatus {
+            slot: self.0,
+            confirmations: Some((current_slot - self.0) as usize),
+            status: match self.3.clone() {
+                Some(err) => Err(err),
+                None => Ok(()),
+            },
+            err: self.3.clone(),
+            confirmation_status: Some(TransactionConfirmationStatus::Finalized),
+        }
+    }
 }
 
-impl Default for RpcConfig {
-    fn default() -> Self {
-        Self {
-            remote_rpc_url: DEFAULT_RPC_URL.to_string(),
-            bind_host: "127.0.0.1".to_string(),
-            bind_port: 8899,
+impl Into<EncodedConfirmedTransactionWithStatusMeta> for TransactionWithStatusMeta {
+    fn into(self) -> EncodedConfirmedTransactionWithStatusMeta {
+        let slot = self.0;
+        let tx = self.1;
+        let meta = self.2;
+        let err = self.3;
+        EncodedConfirmedTransactionWithStatusMeta {
+            slot,
+            transaction: EncodedTransactionWithStatusMeta {
+                transaction: EncodedTransaction::Json(UiTransaction {
+                    signatures: tx.signatures.iter().map(|s| s.to_string()).collect(),
+                    message: UiMessage::Raw(UiRawMessage {
+                        header: tx.message.header,
+                        account_keys: tx
+                            .message
+                            .account_keys
+                            .iter()
+                            .map(|pk| pk.to_string())
+                            .collect(),
+                        recent_blockhash: tx.message.recent_blockhash.to_string(),
+                        instructions: tx
+                            .message
+                            .instructions
+                            .iter()
+                            // TODO: use stack height
+                            .map(|ix| UiCompiledInstruction::from(ix, None))
+                            .collect(),
+                        address_table_lookups: None, // TODO: use lookup table
+                    }),
+                }),
+                meta: Some(UiTransactionStatusMeta {
+                    err: err.clone(),
+                    status: match err {
+                        Some(e) => Err(e),
+                        None => Ok(()),
+                    },
+                    fee: 5000 * (tx.signatures.len() as u64), // TODO: fix calculation
+                    pre_balances: vec![],
+                    post_balances: vec![],
+                    inner_instructions: OptionSerializer::Some(
+                        meta.inner_instructions
+                            .iter()
+                            .enumerate()
+                            .map(|(i, ixs)| UiInnerInstructions {
+                                index: i as u8,
+                                instructions: ixs
+                                    .iter()
+                                    .map(|ix| {
+                                        UiInstruction::Compiled(UiCompiledInstruction {
+                                            program_id_index: ix.instruction.program_id_index,
+                                            accounts: ix.instruction.accounts.clone(),
+                                            data: String::from_utf8(ix.instruction.data.clone())
+                                                .unwrap(),
+                                            stack_height: Some(ix.stack_height as u32),
+                                        })
+                                    })
+                                    .collect(),
+                            })
+                            .collect(),
+                    ),
+                    log_messages: OptionSerializer::Some(meta.logs),
+                    pre_token_balances: OptionSerializer::None,
+                    post_token_balances: OptionSerializer::None,
+                    rewards: OptionSerializer::None,
+                    loaded_addresses: OptionSerializer::None,
+                    return_data: OptionSerializer::Some(UiTransactionReturnData {
+                        program_id: meta.return_data.program_id.to_string(),
+                        data: (
+                            BASE64_STANDARD.encode(meta.return_data.data),
+                            UiReturnDataEncoding::Base64,
+                        ),
+                    }),
+                    compute_units_consumed: OptionSerializer::Some(meta.compute_units_consumed),
+                }),
+                version: Some(TransactionVersion::Legacy(
+                    solana_sdk::transaction::Legacy::Legacy,
+                )),
+            },
+            block_time: SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .map(|d| d.as_millis() as i64)
+                .ok(),
         }
     }
 }
